+++ packages/graph-explorer/src/modules/Namespaces/Namespaces.tsx
@@ -1,65 +1,47 @@
-import { useMemo, useState } from "react";
 import {
   Panel,
   PanelContent,
   PanelHeader,
   PanelHeaderActions,
-  PanelHeaderDivider,
   PanelTitle,
-  Select,
+  SidebarTabs,
+  SidebarTabsContent,
+  SidebarTabsList,
+  SidebarTabsTrigger,
 } from "@/components";
-import { useConfiguration } from "@/core";
 import CommonPrefixes from "./CommonPrefixes";
 import GeneratedPrefixes from "./GeneratedPrefixes";
 import UserPrefixes from "./UserPrefixes";
 import { SidebarCloseButton } from "../SidebarCloseButton";
 
 function Namespaces() {
-  const config = useConfiguration();
-  const [nsType, setNsType] = useState("auto");
-
-  const nsOptions = useMemo(() => {
-    const totalCustom =
-      config?.schema?.prefixes?.filter(
-        prefixConfig => prefixConfig.__inferred !== true
-      ).length || 0;
-
-    const totalAuto =
-      config?.schema?.prefixes?.filter(
-        prefixConfig => prefixConfig.__inferred === true
-      ).length || 0;
-
-    return [
-      { label: `Auto-Generated (${totalAuto})`, value: "auto" },
-      { label: `Custom (${totalCustom})`, value: "custom" },
-      { label: "Common", value: "common" },
-    ];
-  }, [config?.schema?.prefixes]);
-
   return (
     <Panel variant="sidebar">
       <PanelHeader>
         <PanelTitle>Namespaces</PanelTitle>
 
         <PanelHeaderActions>
-          <Select
-            aria-label="Namespace type"
-            options={nsOptions}
-            value={nsType}
-            onChange={v => setNsType(v as string)}
-            hideError={true}
-            noMargin={true}
-            size="sm"
-          />
-          <PanelHeaderDivider />
           <SidebarCloseButton />
         </PanelHeaderActions>
       </PanelHeader>
 
-      <PanelContent className="flex h-full flex-col">
-        {nsType === "auto" && <GeneratedPrefixes />}
-        {nsType === "custom" && <UserPrefixes />}
-        {nsType === "common" && <CommonPrefixes />}
+      <PanelContent className="flex h-full flex-col overflow-hidden">
+        <SidebarTabs defaultValue="auto">
+          <SidebarTabsList>
+            <SidebarTabsTrigger value="auto">Auto-Generated</SidebarTabsTrigger>
+            <SidebarTabsTrigger value="custom">Custom</SidebarTabsTrigger>
+            <SidebarTabsTrigger value="common">Common</SidebarTabsTrigger>
+          </SidebarTabsList>
+          <SidebarTabsContent value="auto">
+            <GeneratedPrefixes />
+          </SidebarTabsContent>
+          <SidebarTabsContent value="custom">
+            <UserPrefixes />
+          </SidebarTabsContent>
+          <SidebarTabsContent value="common">
+            <CommonPrefixes />
+          </SidebarTabsContent>
+        </SidebarTabs>
       </PanelContent>
     </Panel>
   );